Easier method of adding default death cards, some other stuff i forget i'm tired
